1. fix the empty msg hash problem by replacing signed type(int,int64) with uint32, because (int,int64) is not rlp encodable 2. define height/round related fields to be uint32 to align with block definition

+++ api/blocks/types.go
@@ -36,8 +36,8 @@ type Block struct {
 	Nonce            uint64             `json:"nonce"`
 }
 type QC struct {
-	QCHeight         uint64 `json:"qcHeight"`
-	QCRound          uint64 `json:"qcRound"`
+	QCHeight         uint32 `json:"qcHeight"`
+	QCRound          uint32 `json:"qcRound"`
 	VoterBitArrayStr string `json:"voterBitArrayStr"`
 	EpochID          uint64 `json:"epochID"`
 	Raw              string `json:"raw"`
